requirements.txt
Flask==3.1.2
mysql-connector-python==9.1.0
Werkzeug==3.1.5

flask-jwt-extended
flask-limiter
flask-sqlalchemy
python-dotenv
bcrypt